Align HuggingFace types and fix crypto buffer usage
