+++ src/services/environmentConfigLoader.ts
@@ -405,17 +405,17 @@ export class EnvironmentConfigLoader {
     return config;
   }
 
-    private async loadConfigFile(filePath: string, format: 'json' | 'yaml' | 'env'): Promise<any> {
-      const rawContent = readFileSync(filePath, 'utf8');
-      const content = typeof rawContent === 'string' ? rawContent : rawContent.toString('utf8');
+  private async loadConfigFile(filePath: string, format: 'json' | 'yaml' | 'env'): Promise<any> {
+    const rawContent = readFileSync(filePath);
+    const content = rawContent.toString('utf8');
 
-      switch (format) {
-        case 'json':
-          return JSON.parse(content);
-      
+    switch (format) {
+      case 'json':
+        return JSON.parse(content);
+
       case 'yaml':
         return yaml.load(content);
-      
+
       case 'env':
         return this.parseEnvFile(content);
